[FIX] TasteNovelGetResponse DTO의 novelRatingCount 타입 변경
from Integer to Long

+++ src/main/java/org/websoso/WSSServer/dto/userNovel/TasteNovelGetResponse.java
@@ -10,7 +10,7 @@ public record TasteNovelGetResponse(
         String novelImage,
         Long interestCount,
         Float novelRating,
-        Integer novelRatingCount
+        Long novelRatingCount
 ) {
 
     public static TasteNovelGetResponse of(Novel tasteNovel) {
